unpack_test.go,tarutil_test.go: move unpack tests to unpack_test.go
